requirements.txt
Django==5.1.7
djangorestframework==3.14.0
django-cors-headers==4.7.0
django-filter==22.1
psycopg2==2.9.10
sqlparse==0.4.3
requests==2.32.3
urllib3==2.2.3
python-dotenv==0.21.0
fuzzywuzzy python-Levenshtein